src/ai_engine/requirements_313.txt
# AI Engine Python 3.13 Dependencies
# =====================================
# Install in the isolated Python 3.13 environment ONLY.
# DO NOT install these in the main Python 3.14 environment.

# NVIDIA Alchemi (requires Python 3.13)
# nvalchemi-toolkit  # Uncomment when available

# Communication
pyzmq>=25.0.0
msgpack>=1.0.0

# REST fallback
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Scientific computing
numpy>=1.24.0

# PyTorch (for GPU inference)
# torch>=2.0.0  # Uncomment for GPU support

# ASE (Atomic Simulation Environment)
# ase>=3.22.0  # Uncomment when using Alchemi

# HTTP client
requests>=2.31.0
